% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_aic_bic.R
\name{bm_aic_bic}
\alias{bm_aic_bic}
\title{Find best model among all independent variables based on AIC or BIC values.}
\usage{
bm_aic_bic(data, depvar, crit = "AIC")
}
\arguments{
\item{data, depvar, crit:}{data frame containing dependent and independent variables, dependent variable, choose AIC or BIC score.}
}
\value{
A list file containing, best model, its AIC/BIC and adjusted R2
}
\description{
Find best model among all independent variables based on AIC or BIC values.
}
