% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_aic_bic_glm.R
\name{bm_aic_bic_glm}
\alias{bm_aic_bic_glm}
\title{Find best model among all independent variables based on AIC or BIC values.}
\usage{
bm_aic_bic_glm(data, depvar, crit = "AIC", familytype = "gaussian")
}
\arguments{
\item{data, depvar, crit, family}{: data frame containing dependent and independent variables, dependent variable, choose AIC or BIC score, family (for bm_aic_bic_glm only; default: gaussian).}
}
\value{
A list file containing, best model, its AIC/BIC and null deviance
}
\description{
Find best model among all independent variables based on AIC or BIC values.
}
